move esm specific wrap logic to node class
