+++ experimental/packages/opentelemetry-instrumentation/src/instrumentation.ts
@@ -27,7 +27,6 @@ import {
 import * as shimmer from 'shimmer';
 import { InstrumentationModuleDefinition } from './platform/node';
 import * as types from './types';
-import * as util from 'util';
 
 /**
  * Base abstract internal class for instrumenting node and web plugins
@@ -62,50 +61,13 @@ export abstract class InstrumentationAbstract<T = any>
   }
 
   /* Api to wrap instrumented method */
-  protected _wrap = (
-    moduleExports: any,
-    name: any,
-    wrapper: (originalFn: any) => any
-  ) => {
-    if (!util.types.isProxy(moduleExports)) {
-      return shimmer.wrap(moduleExports, name, wrapper);
-    } else {
-      return this._wrapEsm(moduleExports, name, wrapper);
-    }
-  };
+  protected _wrap = shimmer.wrap;
   /* Api to unwrap instrumented methods */
-  protected _unwrap = (moduleExports: any, name: any) => {
-    if (!util.types.isProxy(moduleExports)) {
-      return shimmer.unwrap(moduleExports, name);
-    } else {
-      return this._unwrapEsm(moduleExports, name);
-    }
-  };
+  protected _unwrap = shimmer.unwrap;
   /* Api to mass wrap instrumented method */
   protected _massWrap = shimmer.massWrap;
   /* Api to mass unwrap instrumented methods */
   protected _massUnwrap = shimmer.massUnwrap;
-
-  private _wrapEsm = (
-    moduleExports: T,
-    name: keyof T,
-    wrapper: (original: ({} & T)[keyof T]) => ({} & T)[keyof T]
-  ): void => {
-    const wrapped = shimmer.wrap(
-      Object.assign({}, moduleExports),
-      name,
-      wrapper
-    );
-    Object.defineProperty(moduleExports, name, {
-      value: wrapped,
-    });
-  };
-
-  private _unwrapEsm = (moduleExports: T, name: keyof T): void => {
-    Object.defineProperty(moduleExports, name, {
-      value: moduleExports[name],
-    });
-  };
 
   /* Returns meter */
   protected get meter(): Meter {
